Include files based on `git diff` instead of `grep`
When including untracked files, grep was used to find the tracked source
file in the list of changed files. Now, git-diff is used to detect the
change. It allows to use source files outside of $SYNCROOT.

Source file paths starting with / are seen relative to the Git working
tree, even if --syncroot is specified.